10-10EZ | Clean up constants
