+++ src/applications/hca/components/FormFields/VaMedicalCenter.jsx
@@ -3,14 +3,14 @@ import PropTypes from 'prop-types';
 import { usePrevious } from 'platform/utilities/react-hooks';
 import { apiRequest } from 'platform/utilities/api';
 import { focusElement } from 'platform/utilities/ui';
-import { REACT_BINDINGS, STATES_USA } from '../../utils/imports';
-import { API_ENDPOINTS, STATES_WITHOUT_MEDICAL } from '../../utils/constants';
+import { STATES_USA, VaSelect } from '../../utils/imports';
+import { API_ENDPOINTS } from '../../utils/constants';
 import ServerErrorAlert from '../FormAlerts/ServerErrorAlert';
 import { VaMedicalCenterReviewField } from '../FormReview/VaMedicalCenterReviewField';
 import content from '../../locales/en/content.json';
 
-// expose React binding for web components
-const { VaSelect } = REACT_BINDINGS;
+// declare states without medical care serivces
+const STATES_WITHOUT_MEDICAL = ['AA', 'AE', 'AP', 'FM', 'MH', 'PW'];
 
 const VaMedicalCenter = props => {
   const {
