+++ src/applications/hca/components/FormPages/FinancialInformation.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import FormNavButtons from 'platform/forms-system/src/js/components/FormNavButtons';
-import { LAST_YEAR } from '../../utils/constants';
+import { LAST_YEAR } from '../../utils/helpers';
 
 const HouseholdFinancialOnboarding = props => {
   const {
