+++ src/applications/hca/utils/constants.js
@@ -1,9 +1,3 @@
-import { getAppUrl } from 'platform/utilities/registry-helpers';
-import content from '../locales/en/content.json';
-
-// declare previous year for form questions and content
-export const LAST_YEAR = new Date().getFullYear() - 1;
-
 // declare API endpoint routes
 export const API_ENDPOINTS = {
   csrfCheck: '/maintenance_windows',
@@ -11,21 +5,6 @@ export const API_ENDPOINTS = {
   enrollmentStatus: '/health_care_applications/enrollment_status',
   facilities: '/health_care_applications/facilities',
   ratingInfo: '/health_care_applications/rating_info',
-};
-
-// declare global app URLs for use with content links
-export const APP_URLS = {
-  dischargeWizard: getAppUrl('discharge-upgrade-instructions'),
-  ezr: getAppUrl('ezr'),
-  facilities: getAppUrl('facilities'),
-  hca: getAppUrl('hca'),
-  profile: getAppUrl('profile'),
-  verify: getAppUrl('verify'),
-};
-
-export const DOWNLOAD_ERRORS_BY_CODE = {
-  '5': content['alert-download-message--500'],
-  generic: content['alert-download-message--generic'],
 };
 
 // declare view fields for use in household section
@@ -49,14 +28,14 @@ export const DISABILITY_RATING_INIT_STATE = {
 };
 
 // declare labels for discharge type select box
-export const DISCHARGE_TYPE_LABELS = {
+export const DISCHARGE_TYPE_LABELS = Object.freeze({
   honorable: 'Honorable',
   general: 'General',
   other: 'Other Than Honorable',
   'bad-conduct': 'Bad Conduct',
   dishonorable: 'Dishonorable',
   undesirable: 'Undesirable',
-};
+});
 
 // declare action statuses for fetching enrollment status
 export const ENROLLMENT_STATUS_ACTIONS = {
@@ -139,7 +118,7 @@ export const MOCK_ENROLLMENT_RESPONSE = {
 };
 
 // declare labels for last service branch select box
-export const SERVICE_BRANCH_LABELS = {
+export const SERVICE_BRANCH_LABELS = Object.freeze({
   'air force': 'Air Force',
   army: 'Army',
   'coast guard': 'Coast Guard',
@@ -154,7 +133,7 @@ export const SERVICE_BRANCH_LABELS = {
   'f.scouts new': 'Filipino New Scout',
   'f.scouts old': 'Filipino Old Scout',
   other: 'Other',
-};
+});
 
 // declare name to use for window session storage item
 export const SESSION_ITEM_NAME = 'hcaDependentIndex';
@@ -166,6 +145,3 @@ export const SHARED_PATHS = {
     info: 'household-information/dependent-information',
   },
 };
-
-// declare states without medical care serivces
-export const STATES_WITHOUT_MEDICAL = ['AA', 'AE', 'AP', 'FM', 'MH', 'PW'];
